Optimized CMSIS to use fixed Clock of 30MHz
added #define __USE_DYNAMIC_CLOCK. If enabled the default CMSIS behavior
is observed, where a global variable is used for the clock rate. If
undefined, a fixed value is calculated for the clockrate. To enable
this, the clock definintion was moved to the header file.  This
modification allows to save several hundred bytes.
In addition the clock generation was initialized correctly to use the
internal RC oscillator and yield a 30Mhz clock via the PLL.